+++ lib/src/country_code_picker/functions.dart
@@ -1,22 +1,16 @@
-import 'dart:convert';
-
-import 'package:flutter/material.dart';
+import 'package:flutter_country_code/src/country_code_picker/statics.dart';
 import 'package:flutter_sim_country_code/flutter_sim_country_code.dart';
 
 import 'country.dart';
 
 ///This function returns list of countries
-Future<List<Country>> getCountries(BuildContext context) async {
-  String rawData = await DefaultAssetBundle.of(context)
-      .loadString('packages/flutter_country_code/raw/country_codes.json');
-  final parsed = json.decode(rawData.toString()).cast<Map<String, dynamic>>();
-  return parsed.map<Country>((json) => new Country.fromJson(json)).toList();
-}
+List<Country> getCountries() =>
+    countriesData.map<Country>((json) => new Country.fromJson(json)).toList();
 
 ///This function returns an user's current country. User's sim country code is matched with the ones in the list.
 ///if there's no sim country code it will return `first country on the list`
-Future<Country> getDefaultCountry(BuildContext context) async {
-  final list = await getCountries(context);
+Future<Country> getDefaultCountry() async {
+  final list = getCountries();
   try {
     final countryCode = await FlutterSimCountryCode.simCountryCode;
     if (countryCode == null) {
@@ -30,8 +24,5 @@ Future<Country> getDefaultCountry(BuildContext context) async {
 }
 
 ///This function returns an country whose [countryCode] matches with the passed one.
-Future<Country?> getCountryByCountryCode(
-    BuildContext context, String countryCode) async {
-  final list = await getCountries(context);
-  return list.firstWhere((element) => element.countryCode == countryCode);
-}
+Country getCountryByCountryCode(String countryCode) =>
+    getCountries().firstWhere((element) => element.countryCode == countryCode);